More fixes for Rackspace/OpenCloud
When looking for an object, if it doesn't exist, OpenCloud returns `ObjFetchError`
but the catch for it wasn't catching it without the full namespace...
So I added a catch for the full namespace (and kept the old version too)

Also when writing a file (new file or overwriting)
`$object->bytes` was never being populated...
after digging through the OpenCloud code a bit I couldn't find anywhere,
where `bytes` *would* be set, and without it, our writes never return anything to indicate succes.
So I did a check for if `bytes` empty, but `content_length` is not empty, we return that.