createDirectory return false on sftp.php class
Each time that a new folder is created on the sftp server, mkdir function return false. Nevertheless the folder is created on the sftp server.
The lib php-ssh propose the mkdir functionnality to use directly ssh2_sftp_mkdir.
With this, the function createDirectory return the good value (true) if the folder has been created properly.