Remove purposeless return values from Adapter and Filesystem interfaces
So far, write, delete and rename methods returned values: number of written bytes
for write, and boolean value for the two other methods. This is no longer the case.

The Local adapter was the first adapter to be introduced (see eed10e367),
and presumably the Adapter interface has been based on it. Hence others adapters
try to mimic low-level file functions. In the end it looks like a bad decision:
most adapters are for object storages, not filesystems. Moreover it forces most
adapters to recompute the size of the written content locally.

Considering CQS principle, it's also better to not return any value from
writting methods.