Define env vars in .env file only
The tests are using env vars to run. These env vars are storing the
credentials to the different backends.

For the CI, these env vars are set on the project settings (on travis
for instance).

For the development of Gaufrette, and considering the merge of (#580),
the tests should be ran using docker.
The docker setup uses env vars defined in .env files, so we no longer
need to define such vars in the phpunit.xml file.