:green_heart: Ftp class cannot be spec in PHP 8.1+
This commit removes the test of the class Ftp, it's not possible to spec
it anymore because mocking ftp_* functions does not work anymore. This
is because of the new type \FTP\Connection returned by the function
ftp_open and expected by all other functions. It's not possible anymore
to give any resource to ftp_* functions.