:alien: Upgrade google client library version
First of all, this code is dedicated to provide support for an *old*
SDK of Google. The new one would be much easier to integrate and is
fully compatible with PHP 8.1 while this one trigger a lot of warnings.

However this commit still contains fixes for the new version of
google-api-clients.

It also add a feature that was previously missing: creating bucket if
not exists.