:fire: Remove support for OpenCloud
The package `rackspace/php-opencloud` is not maintained for years, and
will never be. This is why probably nobody use this adapter for now.
It's also important to highlight that openstack has a compatibility
layer with AWS S3 (and we provide great adapters for it).

This is sad, but this is goodbye since even Gaufrette is EOL...